+++ src/04_array_hash/count_trailing_zero_normal.rb
