+++ src/04_array_hash/max_index.rb
