+++ src/04_array_hash/orthogonal_vectors_easy.rb
