+++ src/04_array_hash/pre_pre_last_easy.rb
