+++ src/04_array_hash/prime_factorization_hard.rb
