+++ src/04_array_hash/range_easy.rb
