+++ src/04_array_hash/sort_3_num_easy.rb
