+++ src/04_array_hash/sorted_unique_array_normal.rb
