+++ src/04_array_hash/sum_easy.rb
