+++ src/04_array_hash/validate_isbn10_hard.rb
