--- src/05_exception/errors.rb
@@ -1,5 +1,0 @@
-
-# Define your Errors
-
-# Example
-# class FooError < StandardError; end
