--- src/05_exception/init_array.rb
@@ -1,6 +1,0 @@
-require_relative 'errors'
-
-# return an array except the last one
-def init(arr)
-  arr[1..-1]
-end
