--- src/05_exception/square_non_negative.rb
@@ -1,3 +1,0 @@
-def square_non_negative(num)
-  num * num
-end
