--- src/05_exception/tail_array.rb
@@ -1,6 +1,0 @@
-require_relative 'errors'
-
-# return an array except the first one
-def tail(arr)
-  arr
-end
