+++ test/04_array_hash/count_unsolved.rb
