+++ test/04_array_hash/orthogonal_vectors_easy_test.rb
@@ -1,5 +1,5 @@
 require 'test/unit'
-require_relative '../../src/04_array/orthogonal_vectors_easy'
+require_relative '../../src/04_array_hash/orthogonal_vectors_easy'
 
 class TypeTest < Test::Unit::TestCase
   def test_orthogonal_vectors_cond1
