+++ test/04_array_hash/partial_test.rb
