+++ test/04_array_hash/pre_pre_last_easy_test.rb
@@ -1,5 +1,5 @@
 require 'test/unit'
-require_relative '../../src/04_array/pre_pre_last_easy'
+require_relative '../../src/04_array_hash/pre_pre_last_easy'
 
 class ArrayTest < Test::Unit::TestCase
   def test_pre_pre_last
